# Conflicts:
#	src/main/java/com/opus/opus/modules/team/application/TeamCommandService.java
#	src/main/java/com/opus/opus/modules/team/application/convenience/TeamLikeConvenience.java
#	src/main/java/com/opus/opus/modules/team/domain/dao/TeamRepository.java
#	src/test/java/com/opus/opus/file/FileFixture.java
#	src/test/java/com/opus/opus/team/application/TeamCommandServiceTest.java
#	src/test/java/com/opus/opus/team/application/TeamQueryServiceTest.java

+++ src/main/java/com/opus/opus/modules/team/application/TeamCommandService.java
@@ -30,6 +30,7 @@ import com.opus.opus.modules.file.exception.FileException;
 import com.opus.opus.modules.member.domain.Member;
 import com.opus.opus.modules.team.application.convenience.TeamConvenience;
 import com.opus.opus.modules.team.application.convenience.TeamMemberConvenience;
+import com.opus.opus.modules.team.application.convenience.TeamMemberConvenience;
 import com.opus.opus.modules.team.application.dto.request.TeamCreateRequest;
 import com.opus.opus.modules.team.application.dto.request.TeamUpdateRequest;
 import com.opus.opus.modules.team.application.dto.response.TeamCreateResponse;
@@ -39,8 +40,10 @@ import com.opus.opus.modules.team.domain.Team;
 import com.opus.opus.modules.team.domain.TeamLike;
 import com.opus.opus.modules.team.domain.TeamVote;
 import com.opus.opus.modules.team.domain.dao.TeamLikeRepository;
+import com.opus.opus.modules.team.domain.dao.TeamLikeRepository;
 import com.opus.opus.modules.team.domain.dao.TeamRepository;
 import com.opus.opus.modules.team.domain.dao.TeamVoteRepository;
+import com.opus.opus.modules.team.exception.TeamLikeException;
 import com.opus.opus.modules.team.exception.TeamException;
 import com.opus.opus.modules.team.exception.TeamLikeException;
 import com.opus.opus.modules.team.exception.TeamVoteException;
@@ -61,6 +64,7 @@ public class TeamCommandService {
     private final FileStorageUtil fileStorageUtil;
 
     private final TeamConvenience teamConvenience;
+    private final TeamMemberConvenience teamMemberConvenience;
     private final ContestConvenience contestConvenience;
     private final ContestTrackConvenience contestTrackConvenience;
 
@@ -176,42 +180,52 @@ public class TeamCommandService {
         }
     }
 
-    public void savePreviewImages(final Long teamId, final List<MultipartFile> images) {
+    public void savePreviewImages(final Long teamId, final List<MultipartFile> images, final Member member) {
         teamConvenience.validateExistTeam(teamId);
+        teamMemberConvenience.validateTeamMemberUnlessAdmin(teamId, member);
         checkPreviewLimit(teamId, images);
         for (MultipartFile image : images) {
             fileStorageUtil.storeFile(image, teamId, TEAM, PREVIEW);
         }
     }
 
-    public void deletePreviewImages(final Long teamId, final List<Long> ids) {
+    public void deletePreviewImages(final Long teamId, final List<Long> ids, final Member member) {
         teamConvenience.validateExistTeam(teamId);
+        teamMemberConvenience.validateTeamMemberUnlessAdmin(teamId, member);
         ids.forEach(fileStorageUtil::deleteFile);
     }
 
-    public void saveThumbnailImage(final Long teamId, final MultipartFile image) {
+    public void saveThumbnailImage(final Long teamId, final MultipartFile image, final Member member) {
         teamConvenience.validateExistTeam(teamId);
+        teamMemberConvenience.validateTeamMemberUnlessAdmin(teamId, member);
+        final Optional<File> existingFile = fileRepository.findByReferenceIdAndReferenceTypeAndImageType(teamId, TEAM,
+                THUMBNAIL);
         final Optional<File> existingFile = fileRepository.findByReferenceIdAndReferenceTypeAndImageType(teamId, TEAM,
                 THUMBNAIL);
         fileStorageUtil.storeFile(image, teamId, TEAM, THUMBNAIL);
         existingFile.ifPresent(file -> fileStorageUtil.deleteFile(file.getId()));
     }
 
-    public void deleteThumbnailImage(final Long teamId) {
+    public void deleteThumbnailImage(final Long teamId, final Member member) {
         teamConvenience.validateExistTeam(teamId);
+        teamMemberConvenience.validateTeamMemberUnlessAdmin(teamId, member);
         deleteIfExists(teamId, THUMBNAIL);
     }
 
-    public void savePosterImage(final Long teamId, final MultipartFile image) {
+    public void savePosterImage(final Long teamId, final MultipartFile image, final Member member) {
         teamConvenience.validateExistTeam(teamId);
+        teamMemberConvenience.validateTeamMemberUnlessAdmin(teamId, member);
+        final Optional<File> existingFile = fileRepository.findByReferenceIdAndReferenceTypeAndImageType(teamId, TEAM,
+                POSTER);
         final Optional<File> existingFile = fileRepository.findByReferenceIdAndReferenceTypeAndImageType(teamId, TEAM,
                 POSTER);
         fileStorageUtil.storeFile(image, teamId, TEAM, POSTER);
         existingFile.ifPresent(file -> fileStorageUtil.deleteFile(file.getId()));
     }
 
-    public void deletePosterImage(final Long teamId) {
+    public void deletePosterImage(final Long teamId, final Member member) {
         teamConvenience.validateExistTeam(teamId);
+        teamMemberConvenience.validateTeamMemberUnlessAdmin(teamId, member);
         deleteIfExists(teamId, POSTER);
     }
